+++ src/ios/KalmanFilter.h
@@ -7,15 +7,15 @@
 //
 
 typedef struct {
-    float q;
-    float r;
-    float p;
-    float x;
-    float k;
+    double q;
+    double r;
+    double p;
+    double x;
+    double k;
 } KalmanFilter;
 
-KalmanFilter KalmanFilter_initialise(float processNoise, float measurementNoise, float estimateValue, float estimateVariance);
+KalmanFilter KalmanFilter_initialise(double processNoise, double measurementNoise, double estimateValue, double estimateVariance);
 
-void KalmanFilter_update(float sample, KalmanFilter *kalmanFilter);
+void KalmanFilter_update(double sample, KalmanFilter *kalmanFilter);
 
-float KalmanFilter_estimate(KalmanFilter *kalmanFilter);
+double KalmanFilter_estimate(KalmanFilter *kalmanFilter);
